+++ lib/cool_ui.dart
@@ -4,5 +4,5 @@ import 'package:flutter/material.dart';
 part 'utils/screen_utils.dart';
 part 'utils/widget_utils.dart';
 
-part 'popover/cupertino_popover.dart';
-part 'popover/cupertino_popover_menu_item.dart';
+part 'widgets/popover/cupertino_popover.dart';
+part 'widgets/popover/cupertino_popover_menu_item.dart';