+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 * TODO: 添加了CupertinoPopoverMenuList
 * TODO: 添加了CupertinoPopoverMenuItem
 * TODO: 修改了CupertionPopover动画,
+
+## [0.1.0] - TODO: 改进了CupertionPopover和添加了CupertinoPopoverMenuItem
+
+* TODO: 改进了CupertionPopover箭头的位置
+* TODO: 改进了CupertinoPopoverMenuItem按下的动画，并且添加了onTap